+++ src/cnvyw_.py
@@ -25,6 +25,7 @@ from pywriter.odt.odt_brief_synopsis import OdtBriefSynopsis
 from pywriter.odt.odt_xref import OdtXref
 from pywriter.ods.ods_scenelist import OdsSceneList
 from pywriter.odt.odt_notes import OdtNotes
+from pywriter.odt.odt_todo import OdtTodo
 from pywriter.ods.ods_charlist import OdsCharList
 from pywriter.ods.ods_loclist import OdsLocList
 from pywriter.ods.ods_itemlist import OdsItemList
@@ -177,6 +178,12 @@ def get_notes():
     '''Import Notes chapters from yWriter 7 to a Writer document.
     '''
     open_yw7(OdtNotes.SUFFIX, OdtNotes.EXTENSION)
+
+
+def get_todo():
+    '''Import Todo chapters from yWriter 7 to a Writer document.
+    '''
+    open_yw7(OdtTodo.SUFFIX, OdtTodo.EXTENSION)
 
 
 def get_charlist():